+++ tests/testthat/test-empty.R
@@ -30,9 +30,9 @@ test_that("completely empty sheets are handled [xls]", {
 
 test_that("sheets with column names only are handled", {
   out <- read_excel(test_sheet("empty-sheets.xlsx"), "header_only")
-  expect_identical(out, tibble::tibble(var1 = numeric(), var2 = numeric()))
+  expect_identical(out, tibble::tibble(var1 = logical(), var2 = logical()))
   out <- read_excel(test_sheet("empty-sheets.xls"), "header_only")
-  expect_identical(out, tibble::tibble(var1 = numeric(), var2 = numeric()))
+  expect_identical(out, tibble::tibble(var1 = logical(), var2 = logical()))
 })
 
 test_that("non-empty sheets act that way if we skip past everything", {